+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add table message_categories
 - Add column message_category_id at messages table
 - php artisan update:db
+- create chat_settings table